scripts/external/requirements.txt
# Phase 13 external-data fetcher deps. Pinned conservatively. Do NOT add
# pandas / polars / openmeteo-requests / fetchers SDKs — raw httpx keeps
# the surface uniform across providers.
httpx>=0.27,<1
holidays>=0.25,<1
feedparser>=6.0,<7
PyYAML>=6.0,<7
supabase>=2.0,<3
python-dotenv>=1.0,<2

# Test-only — kept here so a single pip install -r covers CI + local dev.
# pytest is not imported by any production module.
pytest>=8.0,<9
